+++ src/main/java/com/example/app/domain/Withdrawal.java
@@ -8,7 +8,7 @@ import javax.persistence.*;
 @DiscriminatorValue("transaction")
 public class Withdrawal extends Transaction {
 
-	public Withdrawal(User user,Integer amount, LocalDateTime date) {
+	public Withdrawal(User user, Double amount, LocalDateTime date) {
 		super(user, amount, date);
 	}
 }
